+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+### 0.7.4 - 2020-04-27
+
 * Add a Start/Stop the RLS command
 * Introduce a `rust-client.autoStartRls` (defaults to true) setting to control the auto-start
 behaviour when opening a relevant Rust project file